+++ server/component/moduleScheduledJobsCalendar/ModuleScheduledJobsCalendarModel.php
@@ -21,6 +21,11 @@ class ModuleScheduledJobsCalendarModel extends BaseModel
     private $uid;
 
     /**
+     * selected action,
+     */
+    private $action_id;
+
+    /**
      * The constructor.
      *
      * @param array $services
@@ -29,10 +34,11 @@ class ModuleScheduledJobsCalendarModel extends BaseModel
      * @param int $uid
      * The selected user
      */
-    public function __construct($services, $uid)
+    public function __construct($services, $uid, $action_id = -1)
     {
         parent::__construct($services);
         $this->uid = $uid;
+        $this->action_id = $action_id;
     }
 
     /**
@@ -44,6 +50,17 @@ class ModuleScheduledJobsCalendarModel extends BaseModel
     {
         return $this->uid;
     }
+
+    /**
+     * Get selected action
+     * @return int
+     * Return the action id
+     */
+    public function get_selected_action()
+    {
+        return $this->action_id;
+    }
+
 
     /**
      * Set selected user
@@ -63,13 +80,34 @@ class ModuleScheduledJobsCalendarModel extends BaseModel
     public function get_users()
     {
         $arr = array();
-        $sql = "SELECT id, email, code, name 
+        $sql = "SELECT id, email, code, `name` 
                 FROM users u 
-                LEFT JOIN validation_codes c on (c.id_users = u.id)
+                LEFT JOIN validation_codes c ON (c.id_users = u.id)
                 WHERE id_status = :active_status";
         $users = $this->db->query_db($sql, array(':active_status' => USER_STATUS_ACTIVE));
         foreach ($users as $val) {
             array_push($arr, array("value" => intval($val['id']), "text" => ("[" . $val['code'] . '] ' . $val['email']) . ' - ' . $val['name']));
+        }
+        return $arr;
+    }
+
+    /**
+     * Get all actions;
+     * @return array
+     * array used for select dropdown
+     */
+    public function get_actions()
+    {
+        $arr = array();
+        $sql = "SELECT id, `name` 
+                FROM formActions";
+        $actions = $this->db->query_db($sql);
+        $arr[] = array(
+            "value" => -1,
+            "text" => 'All actions'
+        );
+        foreach ($actions as $val) {
+            array_push($arr, array("value" => intval($val['id']), "text" => $val['name']));
         }
         return $arr;
     }
@@ -88,20 +126,23 @@ class ModuleScheduledJobsCalendarModel extends BaseModel
                     CASE 
                         WHEN l_types.lookup_code = 'task' THEN JSON_UNQUOTE(JSON_EXTRACT(t.config, '$.type'))
                         ELSE l_types.lookup_code 
-                    END AS type_code
-                    , `description` as title,
-                    date_to_be_executed AS `start`, DATE_ADD(date_to_be_executed, INTERVAL 1 SECOND) AS `end`, :sjid AS `url`
+                    END AS type_code,
+                    `description` as title,
+                    date_to_be_executed AS `start`, DATE_ADD(date_to_be_executed, INTERVAL 1 SECOND) AS `end`, :sjid AS `url`,
+                    a.id_formActions
                     FROM scheduledJobs sj
                     INNER JOIN scheduledJobs_users sju ON (sj.id = sju.id_scheduledJobs)
                     INNER JOIN lookups l_status ON (l_status.id = sj.id_jobStatus)
                     INNER JOIN lookups l_types ON (l_types.id = sj.id_jobTypes)
                     LEFT JOIN scheduledJobs_tasks sjt ON (sjt.id_scheduledJobs = sj.id)
                     LEFT JOIN tasks t ON (sjt.id_tasks = t.id)
-                    WHERE sju.id_users = :uid";
+                    LEFT JOIN scheduledJobs_formActions a on (a.id_scheduledJobs = sj.id)
+                    WHERE sju.id_users = :uid AND (a.id_formActions = :id_formActions OR :id_formActions = -1)";
             return $this->db->query_db($sql, array(
                 ':uid' => $this->uid,
-                ":sjid" => $scheduled_job_url
-                ));
+                ":sjid" => $scheduled_job_url,
+                ":id_formActions" => $this->action_id
+            ));
         } else {
             return array();
         }