+++ server/db/update_scripts/17_update_v6.2.0_v6.3.0.sql
@@ -11,3 +11,7 @@ SET version = 'v6.3.0';
 -- add title to the page calendar view 
 INSERT IGNORE INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`, `content`) VALUES ((SELECT id FROM pages WHERE keyword = 'moduleScheduledJobsCalendar'), get_field_id('title'), '0000000001', 'Scheduled Jobs - Calendar View');
 INSERT IGNORE INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`, `content`) VALUES ((SELECT id FROM pages WHERE keyword = 'moduleScheduledJobsCalendar'), get_field_id('title'), '0000000002', 'Scheduled Jobs - Calendar View');
+
+UPDATE pages
+SET url = "/admin/scheduledJobs/calendar/[i:uid]?/[i:aid]?"
+WHERE keyword = 'moduleScheduledJobsCalendar'