+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - improve the drag and drop of elements in the CMS page
  - add new field type `panel`. In this field we can add extra elements in customization mode
  - #353 - adding transactions info to scheduled job id page
+ - #352 - filtering actions on calendar view
 
 ### Bugfix
 - bugfix the live search in the dropdown for selecting groups for job actions from type `add_group` or `remove_group`;